finish second step
