+++ packages/rocketchat-integrations/client/stylesheets/integrations.css
@@ -52,6 +52,7 @@
 		list-style: none;
 	}
 }
+
 .integrate-other-ways {
 	& p {
 		font-size: 1rem;
