+++ packages/rocketchat_theme/client/imports/components/header.css
@@ -1,31 +1,15 @@
-.rc-header.rc-header--fullpage {
-	padding: 0;
-
-	& .rc-header__wrap {
-
-		padding: 0;
-
-		box-shadow: none;
-
-		height: 60px
-	}
-
-	& .rc-header__block {
-		font-size: 1.375rem;
-		padding: 0 0.5rem;
-		margin: 0 -0.25rem;
-	}
-
-	& .rc-button {
-		height: 36px;
-		min-height: 36px;
-
-		margin: 0 0.25rem;
-	}
-}
-
 .rc-header {
-	padding: 0 0.5rem;
+	padding: 0.5rem;
+
+	font-size: var(--text-heading-size);
+
+	&--room {
+		padding: 1.25rem;
+
+		box-shadow: 0 1px 2px 0 rgba(31, 35, 41, 0.08);
+
+		font-size: var(--header-title-font-size);
+	}
 
 	&__wrap {
 		z-index: 2;
@@ -34,16 +18,13 @@
 
 		flex: 0 0 auto;
 
+		height: 36px;
+
 		margin: 0 -0.5rem;
 
-		padding: var(--header-padding);
-
-		white-space: nowrap;;
+		white-space: nowrap;
 
 		background-color: var(--header-background-color);
-		box-shadow: 0 1px 2px 0 rgba(31, 35, 41, 0.08);
-
-		font-size: var(--header-title-font-size);
 
 		align-items: center;
 
@@ -53,7 +34,10 @@
 	&__block {
 		display: flex;
 
-		margin: 0 0.5rem;
+		margin: 0 -0.5rem;
+
+		padding: 0 0.5rem;
+
 		align-items: center;
 	}
 
@@ -188,9 +172,10 @@
 	}
 
 	&__toggle-favorite {
-		color: var(--header-toggle-favorite-star-color);
 
-		font-size: 1.5rem;
+		padding: 0 0.25rem;
+
+		color: var(--header-toggle-favorite-star-color);
 
 		&.empty {
 			color: var(--header-toggle-favorite-color);
@@ -201,7 +186,7 @@
 		}
 
 		& > .rc-header__icon {
-			font-size: 1.5rem;
+			font-size: 2rem;
 		}
 
 		&:hover {
@@ -216,8 +201,14 @@
 	&__image {
 		width: 32px;
 		height: 32px;
-		margin: 0 0.5rem;
 		flex-shrink: 0;
+	}
+
+	.rc-button {
+		height: 36px;
+		min-height: 36px;
+
+		margin: 0 0.25rem;
 	}
 }
