+++ packages/rocketchat_theme/client/imports/components/setup-wizard.css
@@ -452,7 +452,7 @@
 			display: flex;
 			flex-direction: row;
 
-			margin: 0 -0.5rem 2rem -0.5rem;
+			margin: 0 -0.5rem 2rem;
 
 			& .rc-button {
 				margin: 0 0.5rem;
