+++ packages/rocketchat_theme/client/imports/general/base.css
@@ -109,6 +109,7 @@ button {
 
 	& .tab-button-icon {
 		color: var(--rc-color-primary-dark);
+
 		font-size: 20px;
 
 		&--star {
